move chunked feature generation
